pyproject.toml
[project]
name = "sec-mcp"
version = "0.1.0"
description = "Deploying an MCP server on Cloud Run"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=3.0.0rc1",
    "google-api-python-client>=2.190.0",
    "google-auth>=2.48.0",
    "google-auth-oauthlib>=1.2.4",
    "google-cloud-storage>=3.9.0",
]
